Reorganize UI package
