+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MapActivity"
+            android:name=".ui.map.MapActivity"
             android:label="@string/app_name"
             android:windowSoftInputMode="adjustResize">
