--- app/src/main/java/com/conorodonnell/bus/Formatting.kt
@@ -1,12 +1,0 @@
-package com.conorodonnell.bus
-
-import com.conorodonnell.bus.api.RealTimeBusInfo
-
-
-fun RealTimeBusInfo.formatBusInfo() = "$route to $destination | ${formatDueTime()}"
-
-private fun RealTimeBusInfo.formatDueTime() =
-    when (duetime) {
-      "Due" -> duetime
-      else -> "$duetime mins"
-    }
