+++ app/src/main/res/layout/activity_map.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.conorodonnell.bus.MapActivity">
+    tools:context="com.conorodonnell.bus.ui.map.MapActivity">
 
     <com.google.android.gms.maps.MapView
         android:id="@+id/mapView"
